Rename searh module names to just be their respective tool name
- searxng_search.py changed to `searxng.py`
- semantic_scholar_search to semantic_scholar
